converting numbers and string part 2

+++ app.py
@@ -1,5 +1,5 @@
 num1 = input("Enter the first number: ")
 num2 = input("Enter the second number: ")
 
-num = int(num1) + int(num2)
+num = float(num1) + float(num2)
 print("Result = " + str(num))
